Making serialization ID's short instead of int. Tested this fix on the customSerialization branch and no test failures occured because of this.
git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@22631 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-java/x10/core/Short.java
@@ -28,7 +28,7 @@ final public class Short extends Number implements StructI, java.lang.Comparable
     x10.lang.Arithmetic<Short>, x10.lang.Bitwise<Short>, x10.util.Ordered<Short>
 {
     private static final long serialVersionUID = 1L;
-    private static final int _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(Short.class);
+    private static final short _serialization_id = x10.x10rt.DeserializationDispatcher.addDispatcher(Short.class);
     
     public static final x10.rtt.RuntimeType<?> $RTT = Types.SHORT;
     public x10.rtt.RuntimeType<?> $getRTT() {return $RTT;}
@@ -166,7 +166,7 @@ final public class Short extends Number implements StructI, java.lang.Comparable
         serializer.write($value);
     }
 
-    public int $_get_serialization_id() {
+    public short $_get_serialization_id() {
         return _serialization_id;
     }
